fix: inability to install extension
Newtonsoft.json dll is explicitly ignored by the Playnite Toolbox utility which is used to pack the extension. For this reason, packaged extensions are installed without this dll and throw an error when Playnite is run.

Although Playnite SDK provides a serialization utility, it does not allow configuration of casing (camel case). For this reason, System.Text.Json is used as Newtonsoft's replacement.

Fixes #230