wip: releases' platform
Derive a release's platform from the game release's source and available platforms. In the database, every game (all its releases) individual releases contains a collection of all available platforms. Example:

In other words, Elden Ring has two releases. Both have a source of PlayStation. Both have a platforms collection of PS4, PS5, PC, etc. From this we expect to get a single game, which has 2 releases, each with a single platform of PS4 and one of PS5. This is because the release's source is PlayStation. Each release should condense its collection of platforms to a single value, based on the source. Some sources, will have multiple platforms, e.g. PlayStation has platforms PS3, PS4, PS5. There should be a release corresponding to each platform for a given source.

Note, the types provided by the database and the types that resolvers returned were previously the same. However, given we need to distill down information derived from many database items, the types have diverged. Example, the database has a source and collection of platforms for a release. However, we resolve a release to contain a single platform derived from its source and platforms. For this reason, the previous types have been renamed to include `Db` in the name and a new `resolverTypes` contains the types for the resolvers.