querying data does not guarantee order
When querying for all __ with a given Id from a collection of Ids, there is not guarantee the results will be in the same order as the input Ids. For this reason, we must order the results to be in the same order as the incoming Ids. Otherwise, an incorrect result is associated with each Id.