feat(ui): 🎸 CSP is customizable
CSP, or the content security policy, may now be customized via an
environment variable, `CSP_ORIGINS`. This value may be one or more
(comma-delimited string) of origins that will be allowed to load images,
styles, fonts, and scripts. Note that an allowed origin may load any of
the above mentioned types of asset.

✅ Closes: #498