fix(api): 🐛 do not fail when querying data contianing binary data
It is possible for game data to contain binary data. In these cases, querying for game data can throw an error if the string with binary data is accessed.

This change will prevent the error from being thrown. This allows the user to see games. This PR does NOT include logging to notify the user which game is problematic to aid in fixing the root cause.

✅ Closes: #522