fix(api): 🐛 activating game requires game and platform IDs
Despite a game release having an ID, it not possible to determine a game
and platform from a game release ID alone. This is because the game
(set of releases) is mapped to a platform in a priority order;
e.g. PS5 before Ps4. There is not any actual data on the game release
from Playnite to uniquely identify a release to a specific platform.

BREAKING CHANGE: 🧨 activate game now requires a game ID and platform ID instead of a game
release ID.

✅ Closes: #543