chore(updater): sync game vs release
This change introduces the concept of a game and its releases into the PlayniteWebPlugin.

- Published games and releases include the correct platform.
  - For PC, this is accomplished by comparing the source name to PC platforms.
  - For other platforms, the platforms matching the source are sorted by most relevant and assigned to a release this way.

BREAKING CHANGE: You MUST drop your database prior to upgrading to this version. Previously, releases were stored in the game collection. This is because Playnite does not have a concept of a game versus its releases. This change introduces the concept of a game and its releases into the PlayniteWebPlugin. Therefore, the game collection is now for games and there is a new collection for releases.