fix: adding a release in Playnite does not corrupt database
When adding a new game to the Playnite library then the release is properly published to Playnite Web. Previously, the message was being publsihed as a game and not a release.

Additionally, when a Playnite game is removed, the release removal is published instead of a Game removal. If the removed Playnite Game is the last release for a Game in Playnite Web, then the game removal is also published.