wip: update messages are persistent
MQTT messages are not structured to be persistent for entity updates. This ensures updates applied in Playnite Web or Playnite are received to keep the two in sync, even if one is offline. Additionally, the device ID becomes important to prevent infinite loop updates of Playnite or Playnite  Web receiving update messages that it itself broadcast.

Note the MQTT documentation for the changes to message formats.

BREAKING CHANGE